HSEARCH-3787 Move TimingSource to spi
