+++ engine/src/main/java/org/hibernate/search/engine/backend/spi/BackendBuildContext.java
@@ -6,7 +6,7 @@
  */
 package org.hibernate.search.engine.backend.spi;
 
-import org.hibernate.search.engine.common.timing.impl.TimingSource;
+import org.hibernate.search.engine.common.timing.spi.TimingSource;
 import org.hibernate.search.engine.environment.thread.spi.ThreadPoolProvider;
 import org.hibernate.search.engine.reporting.FailureHandler;
 import org.hibernate.search.engine.environment.bean.BeanResolver;
