+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/search/loading/impl/HibernateOrmLoadingSessionContext.java
@@ -8,7 +8,7 @@ package org.hibernate.search.mapper.orm.search.loading.impl;
 
 import org.hibernate.engine.spi.SessionImplementor;
 import org.hibernate.search.engine.backend.common.spi.DocumentReferenceConverter;
-import org.hibernate.search.engine.common.timing.impl.TimingSource;
+import org.hibernate.search.engine.common.timing.spi.TimingSource;
 import org.hibernate.search.mapper.orm.common.EntityReference;
 
 public interface HibernateOrmLoadingSessionContext {
